move remoteOperations to adapter
