+++ packages/devextreme/js/__internal/grids/grid_core/data_source_adapter/utils/store.ts
